pyproject.toml
[tool.poetry]
name = "sol-cgt"
version = "0.1.0"
description = "Solana capital gains tooling"
authors = ["Auto Dev <auto@example.com>"]
packages = [{ include = "sol_cgt" }]

[tool.poetry.dependencies]
python = "^3.11"
httpx = {version = "^0.27.0", extras = ["http2"]}
pydantic = "^2.5"
pydantic-settings = "^2.1"
rich = "^13.7"
pandas = "^2.1"
pyarrow = "^14.0"
python-dateutil = "^2.9"
tzdata = "^2023.3"
typer = "^0.9"
orjson = "^3.9"
cachetools = "^5.3"
tenacity = "^8.2"
numpy = "^1.26"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1"
pytest-asyncio = "^0.23"
freezegun = "^1.4"

[tool.poetry.scripts]
solcgt = "sol_cgt.cli:app"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["sol_cgt/tests"]

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
